refact: chat template saved on file
